Add files via upload
This notebook contains Exploratory Data Analysis (EDA) on a COVID-19 dataset using Python. It includes data cleaning, visualization, and statistical analysis to uncover insights about the spread and impact of the virus. Libraries such as Pandas, Matplotlib, and Seaborn are used to explore trends, missing values, and correlations in the dataset.